Move `configurationAttributesSchema` down to `DependencyHandler`
This commit fixes gradle/performance#251 by moving the attributes schema to the dependency handler. It's worth noting that
is also renames the configuration method from `configurationAttributesSchema` to simply `attributesSchema`.
