+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/ArtifactDeclarationIntegrationTest.groovy
@@ -22,8 +22,10 @@ class ArtifactDeclarationIntegrationTest extends AbstractIntegrationSpec {
     def setup() {
         buildFile << """
             allprojects {
-                configurationAttributesSchema {
-                    attribute(Attribute.of('usage', String))
+                dependencies {
+                    attributesSchema {
+                        attribute(Attribute.of('usage', String))
+                    }
                 }
                 configurations { compile { attributes usage: 'for-compile' } }
             }
