+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/ResolvedArtifactsApiIntegrationTest.groovy
@@ -28,8 +28,10 @@ rootProject.name = 'test'
 """
         buildFile << """
 allprojects {
-    configurationAttributesSchema {
-       attribute(Attribute.of('usage', String))
+    dependencies {
+       attributesSchema {
+          attribute(Attribute.of('usage', String))
+       }
     }
     configurations {
         compile {
