+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/ResolvedFilesApiIntegrationTest.groovy
@@ -26,8 +26,10 @@ rootProject.name = 'test'
 """
         buildFile << """
 allprojects {
-    configurationAttributesSchema {
-       attribute(Attribute.of('usage', String))
+    dependencies {
+        attributesSchema {
+           attribute(Attribute.of('usage', String))
+        }
     }
     configurations {
         compile {
